sql_app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
pillow
numpy
google-generativeai
google-ai-generativelanguage
ipython
textwrap
python-dotenv
transformers
datasets
torch
pandas
pytorch
tensorflow
flax